+++ R/sigmajs.R
@@ -38,9 +38,6 @@ sigmajs <- function(type = "canvas", width = "100%", kill = FALSE, height = NULL
 		  crosstalk_group = NULL
 		)
   )
-  
-  if(rstudioapi::isAvailable() && isTRUE(interactive()))
-    warning("Graph does not show in the RStudio Viewer; opening in browser", call. = FALSE)
 
   # create widget
   htmlwidgets::createWidget(
@@ -51,10 +48,8 @@ sigmajs <- function(type = "canvas", width = "100%", kill = FALSE, height = NULL
     package = 'sigmajs',
     elementId = elementId,
     sizingPolicy = htmlwidgets::sizingPolicy(
-      viewer.suppress = TRUE,
       browser.fill = TRUE,
-      padding = 20,
-      browser.external = TRUE
+      padding = 20
     ),
     dependencies = crosstalk::crosstalkLibs()
   )